pyproject.toml
[tool.poetry]
name = "velib-forecasting"
version = "0.1.0"
description = "Package for the Velib forecasting project"
authors = ["HippolyteGuigon <hippolyte.guigon@polytechnique.edu>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas-gbq = "^0.22.0"
google-auth = "^2.29.0"
pandera = "^0.18.3"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
